gimlet/cosmo: Put back `U1` in TMP117 refdes (#2204)
In [this comment][1] on PR #2199, @mkeeter suggested we not add `/U1` to
the refdes for TMP117 temperature sensors., and just use the `Jxxx`
jack. I followed Matt's advice because I didn't remember why I had added
that. It turns out I shouldn't have, because `host-sp-comms` was
actually adding it explicitly.

The presence of the `/U1` is apparently necessary for `fmtopo` to
understand the temperature sensors on the host side, and they're not
present with it removed (as determined by `diff`ing the `fmtopo -V`
output on `master` vs R16). Thus, this branch undoes the consequences of
listening to Matt, and puts it back. Now, the temperature sensors once
again show up in `fmtopo -V`

[1]: https://github.com/oxidecomputer/hubris/pull/2199#discussion_r2291387093