`minicbor-lease` adapter for encoding straight into a lease (#2164)
Based on [a suggestion][1] from @mkeeter, this branch adds an adapter
that implements `minicbor::encode::write::Write` for a cursor into a
writable lease. This lets us avoid double-buffering in the
`read_ereports` path, where we would previously encode each ereport into
the receive buffer, and then copy the contents of that into the lease if
there is space for it. The new approach avoids the memcpy, and also
doesn't encode the ereport in the case where there isn't space left for
it.

I implemented the lease-writer adapter thing in a separate
`minicbor-lease` crate, as I was hoping it would be useful
elsewhere...but I'm not actually sure if it will be, now that I think of
it: the IPC for delivering an ereport to packrat leases the data _from_
the caller, so callers will just encode into a normal buffer they own.
Ah well.

[1]:
https://github.com/oxidecomputer/hubris/pull/2126#discussion_r2198529251