cpu_seq: standardize ereport naming (#2231)
This commit tweaks the field and class naming in the CPU sequencer
ereports for VRM PMbus alerts (added in #2184) based on a discussion
with @rmustacc on standardizing ereport naming and class hierarchy. In
particular, we namespace all "hardware" events under "hw", and emit a
different class name string when the POWER_GOOD state has changed.
Additionally, we've rename the "status" field for PMBus status registers
to "pmbus_status" for clarity, and the "dev_id" to "refdes" --- the idea
is that "dev_id" will be used when we want to indicate that the
`control-plane-agent` device ID is _different_ from the refdes, but all
`hw.` ereports will include a "refdes".